etl/src/requirements.txt
aioredis==2.0.1
backoff==2.2.1
elasticsearch==7.9.1
pydantic==1.10.4
python-dotenv==0.21.1
psycopg2-binary==2.9.5
retry==0.9.2